pyproject.toml
[tool.poetry]
name = "course-work-4"
version = "0.1.0"
description = ""
authors = ["Pinchuk Sergey <Pinboyer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
requests = "^2.32.3"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
